Bump Scala, Scalaz and Hedgehog - Scala 2 to 2.13.16 - Scala 3 to 3.1.3 - Scalaz to 7.2.36 - Hedgehog to 0.10.1
Also update the code to fix compilation issue caused by Scala version bumping
