+++ .github/workflows/build.yml
@@ -10,9 +10,9 @@ jobs:
     strategy:
       matrix:
         scala:
-          - { name: "Scala 2", version: "2.12.17", binary-version: "2.12", java-version: "adopt@1.11" }
-          - { name: "Scala 2", version: "2.13.10", binary-version: "2.13", java-version: "adopt@1.11" }
-          - { name: "Scala 3", version: "3.0.2",   binary-version: "3",    java-version: "adopt@1.11" }
+          - { name: "Scala 2", version: "2.12.20", binary-version: "2.12", java-version: "adopt@1.11" }
+          - { name: "Scala 2", version: "2.13.16", binary-version: "2.13", java-version: "adopt@1.11" }
+          - { name: "Scala 3", version: "3.1.3",   binary-version: "3",    java-version: "adopt@1.11" }
 
     steps:
       - uses: actions/checkout@v2.3.4
