--- modules/pirate-scalaz/src/test/scala-2.13-/pirate/spec/LawsCompat.scala
@@ -1,5 +1,0 @@
-package pirate.spec
-
-trait LawsCompat {
-  type LazyCollection[+A] = Stream[A]
-}
