+++ modules/pirate-scalaz/src/test/scala-3/pirate/example/CutExample.scala
