+++ modules/pirate-scalaz/src/test/scala-3/pirate/example/GitExample.scala
