+++ modules/pirate-scalaz/src/test/scala-3/pirate/example/InteractiveExample.scala
