+++ modules/pirate-scalaz/src/test/scala-3/pirate/example/MthExample.scala
